Modified the window increase algorithm
Modified the window increase algorithm from multiplicative increase to additive increase in ConsumerWindow. Initial experiments show that this seems stabler and generates much less NACKs.

As for the window decrease algorithm, the current implementation decreases window size by one for each NACK. This is not multiplicative decrease. Instead, it is similar to the behavior of Relentless Congestion Control (http://staff.psc.edu/mathis/relentless/). At this point, this behavior appears reasonable to me.